Añadiendo enpoint de pago

+++ src/main/java/com/udea/vuelo/model/Price.java
@@ -53,7 +53,7 @@ public class Price {
         this.otherCosts = otherCosts;
     }
 
-    public int getTotalCost() {
+    public long getTotalCost() {
         return totalCost;
     }
